requirements.txt
streamlit>=1.32.0
pandas>=2.2.1
openpyxl>=3.1.2
supabase>=2.3.0
reportlab>=4.1.0
python-docx>=1.1.2
streamlit-sortables>=0.2.0
pytest>=8.1.1
pytest-cov>=5.0.0
chardet>=5.2.0
python-dotenv>=1.0.1
